backend/requirements.txt
# Wizard404 Backend - Dependencias
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# DB
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Auth
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Semantic search (optional: búsqueda por embeddings con all-MiniLM-L6-v2)
sentence-transformers>=2.2.0
numpy>=1.24.0

# Document extraction
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.2
Pillow>=10.0.0

# CLI (compartido con backend para ejecutar wizard404_cli)
typer>=0.9.0
rich>=13.0.0
textual>=0.47.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
